Cleanup
Took out some comments and the given path so that path input is the right path. I changed the path for where to send the files to ..//Networks2//received_files etc to see if we change the directory of the project if the path will still be valid. It needs to be tested on a different computer though.